Change mtvec implementation and boot address
This commit changes the way mtvec is read from and written to in
order to comply with the RISC-V spec allowing for it to be written
and read from by software.
This also means that the boot address can now be set to the input
boot address rather than an offset of 0x80 from it.